// online-drung-detail-web/drung/src/app/models/drung.ts
export class Drung {
  drungId: number;
  drungName: String;
  drungType: {
    drungtypeId: number;
    typeName: String;
  }
	dosage: String;
	propreties: String;
	beforeUse: String;
	howUse: String;
	codeConduct: String;
	danger: String;
	keep: String;
	look: String;
	image: String;
	referUrl: String;
}